Split off buy group from enchanted-armoire command
